Updated files that had misspelled Caps
