+++ src/components/tabNav.scss
@@ -1,3 +1,4 @@
+
 .TabNav {
   display: flex;
   width: 75%;
@@ -32,4 +33,4 @@
   top: 0;
   left: 20;
 
-}
+}
