+++ src/index.js
@@ -3,7 +3,7 @@ import Route from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter as Router } from "react-router-dom";
 import "./index.css";
-import App from "./App";
+import App from "./App"; 
 
 
 import TabNav from "./components/TabNav.js";
